+++ src/Move.h
@@ -1,5 +1,6 @@
 #ifndef CHESSENGINE_MOVE_H
 #define CHESSENGINE_MOVE_H
+#include <cstdint>
 #include <iostream>
 #include "Piece.h"
 #include "Util.h"
@@ -21,8 +22,8 @@ namespace Moves {
 	};
 
 	struct Move {
-		short start;
-		short end;
+		int_fast8_t start;
+		int_fast8_t end;
 		Type promotionType;
 		MoveFlag flag;
 		Color player;
@@ -44,7 +45,7 @@ namespace Moves {
 
 	constexpr Move NULL_MOVE = Move{0, 0, Pieces::EMPTY, QUIET, Color::WHITE};
 
-	constexpr short MAILBOX[120] = {
+	constexpr int_fast8_t MAILBOX[120] = {
 		-1, -1, -1, -1, -1, -1, -1, -1, -1, -1,
 		-1, -1, -1, -1, -1, -1, -1, -1, -1, -1,
 		-1, 0, 1, 2, 3, 4, 5, 6, 7, -1,
@@ -59,7 +60,7 @@ namespace Moves {
 		-1, -1, -1, -1, -1, -1, -1, -1, -1, -1
 	};
 
-	constexpr short MAILBOX64[64] = {
+	constexpr int_fast8_t MAILBOX64[64] = {
 		21, 22, 23, 24, 25, 26, 27, 28,
 		31, 32, 33, 34, 35, 36, 37, 38,
 		41, 42, 43, 44, 45, 46, 47, 48,
@@ -71,8 +72,8 @@ namespace Moves {
 	};
 
 	constexpr bool SLIDE[7] = {false, false, false, true, true, true, false};
-	constexpr short OFFSETS[7] = {0, 4, 8, 4, 4, 8, 8};
-	constexpr short OFFSET[7][8] = {
+	constexpr int_fast8_t OFFSETS[7] = {0, 4, 8, 4, 4, 8, 8};
+	constexpr int_fast8_t OFFSET[7][8] = {
 		{0, 0, 0, 0, 0, 0, 0, 0},
 		{8, 9, 11, 16, 0, 0, 0, 0},
 		{-21, -19, -12, -8, 8, 12, 19, 21},
